move ui directory to elements
