--- src/components/link-github-icon-button/hooks.ts
@@ -1,3 +1,0 @@
-export const useTitle = () => {
-  //
-}
