+++ src/components/link-github-icon-button/link-github-icon-button.tsx
@@ -3,12 +3,12 @@
 import { GitHubLogoIcon } from '@radix-ui/react-icons'
 import Link from 'next/link'
 
-import { Button } from '@/components/root/ui/button/button'
+import { Button } from '@/elements/root/button/button'
 import {
   Tooltip,
   TooltipContent,
   TooltipTrigger
-} from '@/components/root/ui/tooltip/tooltip'
+} from '@/elements/root/tooltip/tooltip'
 
 import { useDictionary } from '@/contexts/root/DictionaryProvider'
